+++ backend/registry/mapping/capabilities/wfs/wfs.py
@@ -1,77 +1,35 @@
-from copy import deepcopy
 from importlib import import_module
 
-from odin.mapping import (MappingBase, MappingMeta, assign, define,
-                          forward_mapping_factory)
-from ows_lib.xml_mapper.capabilities.mixins import OperationUrl
+from odin.mapping import MappingMeta, assign, forward_mapping_factory
 from ows_lib.xml_mapper.utils import get_import_path_for_xml_mapper
-from registry.models.metadata import DatasetMetadata
+from registry.mapping.capabilities.base import (
+    OgcServiceElementToXmlMappingBase, OgcServiceToXmlMappingBase)
 from registry.models.service import FeatureType
 
 
-class FeatureTypeToXml(MappingBase, metaclass=MappingMeta):
-
-    def update(self, destination_obj, *args, **kwargs):
-        self._destination_obj = destination_obj
-        return super().update(destination_obj=self._destination_obj, *args, **kwargs)
-
-    def keywords(self):
-        return [str(keyword) for keyword in self.source.keywords.all()]
-
-    def remote_metadata(self):
-        # FIXME: find style and update instead of creating them with apply
-        xml_metadata_url_mapper_cls = getattr(
-            import_module(
-                get_import_path_for_xml_mapper(
-                    self._destination_obj.serializeDocument()
-                )
-            ),
-            "RemoteMetadata")
-        remote_metadata_mapper_cls = forward_mapping_factory(
-            from_obj=DatasetMetadata, to_obj=xml_metadata_url_mapper_cls, mappings=[define(from_field="origin_url", to_field="link")])
-        _remote_metadata = []
-        for remote_metadata in self.source.dataset_metadata_relations.all():
-            _remote_metadata.append(
-                remote_metadata_mapper_cls.apply(source_obj=remote_metadata))
-        return _remote_metadata
+class FeatureTypeToXmlMappingBase(OgcServiceElementToXmlMappingBase, metaclass=MappingMeta):
+    pass
 
 
-class WebFeatureServiceToXml(MappingBase, metaclass=MappingMeta):
+class WebFeatureServiceToXmlMappingBase(OgcServiceToXmlMappingBase, metaclass=MappingMeta):
 
-    def keywords(self):
-        return [str(keyword) for keyword in self.source.keywords.all()]
-
-    def update(self, destination_obj, *args, **kwargs):
-        self._destination_obj = deepcopy(destination_obj)
-        updated_service = super().update(
-            destination_obj=self._destination_obj, *args, **kwargs)
-        self._update_operation_urls(updated_service=updated_service)
+    def update(self, *args, **kwargs):
+        updated_service = super().update(*args, **kwargs)
         self._update_feature_types()
         return updated_service
-
-    def _update_operation_urls(self, updated_service):
-        operation_urls = []
-        for operation_url in self.source.operation_urls.all():
-            operation_urls.append(
-                OperationUrl(
-                    method=operation_url.method,
-                    url=operation_url.url,
-                    operation=operation_url.operation,
-                    mime_types=[str(mime_type) for mime_type in operation_url.mime_types.all()]))
-        updated_service.operation_urls = operation_urls
 
     def _update_feature_types(self):
         mappings = []
         mappings.append(assign(to_field="remote_metadata",
-                        action=FeatureTypeToXml.remote_metadata, to_list=True))
+                        action=FeatureTypeToXmlMappingBase.remote_metadata, to_list=True))
         mappings.append(assign(to_field="keywords",
-                        action=FeatureTypeToXml.keywords, to_list=True))
+                        action=FeatureTypeToXmlMappingBase.keywords, to_list=True))
 
         for feature_type in self.source.featuretypes.filter(is_active=True):
             xml_feature_type_mapper_cls = getattr(import_module(get_import_path_for_xml_mapper(
                 self._destination_obj.serializeDocument())), "FeatureType")
             mapper_cls = forward_mapping_factory(
-                from_obj=FeatureType, to_obj=xml_feature_type_mapper_cls, base_mapping=FeatureTypeToXml, mappings=mappings)
+                from_obj=FeatureType, to_obj=xml_feature_type_mapper_cls, base_mapping=FeatureTypeToXmlMappingBase, mappings=mappings)
             mapper = mapper_cls(source_obj=feature_type)
             mapper.update(destination_obj=self._destination_obj.get_feature_type_by_identifier(
                 identifier=feature_type.identifier))
