+++ backend/registry/mapping/capabilities/wms/wms.py
@@ -1,22 +1,14 @@
-from copy import deepcopy
 from importlib import import_module
 
-from odin.mapping import (MappingBase, MappingMeta, assign, define,
-                          forward_mapping_factory)
-from ows_lib.xml_mapper.capabilities.mixins import OperationUrl
+from odin.mapping import MappingMeta, assign, forward_mapping_factory
 from ows_lib.xml_mapper.utils import get_import_path_for_xml_mapper
-from registry.models.metadata import DatasetMetadata, Style
+from registry.mapping.capabilities.base import (
+    OgcServiceElementToXmlMappingBase, OgcServiceToXmlMappingBase)
+from registry.models.metadata import Style
 from registry.models.service import Layer
 
 
-class LayerToXml(MappingBase, metaclass=MappingMeta):
-
-    def update(self, destination_obj, *args, **kwargs):
-        self._destination_obj = destination_obj
-        return super().update(destination_obj=self._destination_obj, *args, **kwargs)
-
-    def keywords(self):
-        return [str(keyword) for keyword in self.source.keywords.all()]
+class LayerToXmlMappingBase(OgcServiceElementToXmlMappingBase, metaclass=MappingMeta):
 
     def styles(self):
         # FIXME: find style and update instead of creating them with apply
@@ -29,62 +21,28 @@ class LayerToXml(MappingBase, metaclass=MappingMeta):
             _styles.append(style_mapper_cls.apply(source_obj=style))
         return _styles
 
-    def remote_metadata(self):
-        # FIXME: find style and update instead of creating them with apply
-        xml_metadata_url_mapper_cls = getattr(
-            import_module(
-                get_import_path_for_xml_mapper(
-                    self._destination_obj.serializeDocument()
-                )
-            ),
-            "RemoteMetadata")
-        remote_metadata_mapper_cls = forward_mapping_factory(
-            from_obj=DatasetMetadata, to_obj=xml_metadata_url_mapper_cls, mappings=[define(from_field="origin_url", to_field="link")])
-        _remote_metadata = []
-        for remote_metadata in self.source.dataset_metadata_relations.all():
-            _remote_metadata.append(
-                remote_metadata_mapper_cls.apply(source_obj=remote_metadata))
-        return _remote_metadata
 
+class WebMapServiceToXmlMappingBase(OgcServiceToXmlMappingBase, metaclass=MappingMeta):
 
-class WebMapServiceToXml(MappingBase, metaclass=MappingMeta):
-
-    def keywords(self):
-        return [str(keyword) for keyword in self.source.keywords.all()]
-
-    def update(self, destination_obj, *args, **kwargs):
-        self._destination_obj = deepcopy(destination_obj)
-        updated_service = super().update(
-            destination_obj=self._destination_obj, *args, **kwargs)
-        self._update_operation_urls(updated_service=updated_service)
+    def update(self, *args, **kwargs):
+        updated_service = super().update(*args, **kwargs)
         self._update_layers()
         return updated_service
-
-    def _update_operation_urls(self, updated_service):
-        operation_urls = []
-        for operation_url in self.source.operation_urls.all():
-            operation_urls.append(
-                OperationUrl(
-                    method=operation_url.method,
-                    url=operation_url.url,
-                    operation=operation_url.operation,
-                    mime_types=[str(mime_type) for mime_type in operation_url.mime_types.all()]))
-        updated_service.operation_urls = operation_urls
 
     def _update_layers(self):
         mappings = []
         mappings.append(assign(to_field="remote_metadata",
-                        action=LayerToXml.remote_metadata, to_list=True))
+                        action=LayerToXmlMappingBase.remote_metadata, to_list=True))
         mappings.append(
-            assign(to_field="styles", action=LayerToXml.styles, to_list=True))
+            assign(to_field="styles", action=LayerToXmlMappingBase.styles, to_list=True))
         mappings.append(assign(to_field="keywords",
-                        action=LayerToXml.keywords, to_list=True))
+                        action=LayerToXmlMappingBase.keywords, to_list=True))
 
         for layer in self.source.layers.filter(is_active=True):
             xml_layer_mapper_cls = getattr(import_module(get_import_path_for_xml_mapper(
                 self._destination_obj.serializeDocument())), "Layer")
             layer_mapper_cls = forward_mapping_factory(
-                from_obj=Layer, to_obj=xml_layer_mapper_cls, base_mapping=LayerToXml, mappings=mappings)
+                from_obj=Layer, to_obj=xml_layer_mapper_cls, base_mapping=LayerToXmlMappingBase, mappings=mappings)
             layer_mapper = layer_mapper_cls(source_obj=layer)
             layer_mapper.update(destination_obj=self._destination_obj.get_layer_by_identifier(
                 identifier=layer.identifier))
